refactor: 경로수정, 상수 수정
